chore: separated initoptions, renamed interfaces
