+++ Assets/PlayroomKit/PlayroomKit.cs
@@ -14,47 +14,7 @@ namespace Playroom
     public partial class PlayroomKit
     {
         private static bool isPlayRoomInitialized;
-
         private static readonly Dictionary<string, Player> Players = new();
-
-        [Serializable]
-        public class InitOptions
-        {
-            public bool streamMode = false;
-            public bool allowGamepads = false;
-            public string baseUrl = "";
-            public string[] avatars = null;
-            public string roomCode = "";
-            public bool skipLobby = false;
-            public int reconnectGracePeriod = 0;
-            public int? maxPlayersPerRoom;
-            public string? gameId;
-            public bool discord = false;
-
-            public Dictionary<string, object> defaultStates = null;
-            public Dictionary<string, object> defaultPlayerStates = null;
-
-            private object matchmakingField;
-
-            // Property to handle matchmaking as either boolean or MatchMakingOptions
-            public object matchmaking
-            {
-                get => matchmakingField;
-                set
-                {
-                    if (value is bool || value is MatchMakingOptions)
-                    {
-                        matchmakingField = value;
-                    }
-                    else
-                    {
-                        throw new ArgumentException(
-                            "matchmaking must be either a boolean or a MatchMakingOptions object.");
-                    }
-                }
-            }
-        }
-
 
         public class MatchMakingOptions
         {
@@ -761,7 +721,7 @@ namespace Playroom
         {
             if (Players.TryGetValue(playerId, out Player player))
             {
-                ((IPlayerInteraction)player).InvokeOnQuitWrapperCallback();
+                ((PlayroomKit.IPlayerInteraction)player).InvokeOnQuitWrapperCallback();
             }
             else
             {
