+++ Assets/PlayroomKit/ServiceHandler.cs
@@ -1,0 +1,7 @@
+namespace Playroom
+{
+    public class Playroomkit
+    {
+        // TODO USE THIS 
+    }
+}
